requirements.txt
aiodns==3.1.1
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
blinker==1.7.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==24.1.3
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
cryptography==41.0.7
dbus-python==1.3.2
Deprecated==1.2.14
distro==1.9.0
distro-info==1.7+build1
fastapi==0.101.0
Flask==3.0.2
Flask-Cors==4.0.0
frozenlist==1.4.0
h11==0.14.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.4
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
packaging==24.0
pexpect==4.9.0
psutil==5.9.8
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyasyncore==1.0.2
pycares==4.4.0
pydantic==1.10.14
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
pyinotify==0.9.6
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu3
python-debian==0.1.49+ubuntu2
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==4.3.4
requests==2.31.0
rich==13.7.1
screen-resolution-extra==0.0.0
service-identity==24.1.0
setuptools==68.1.2
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
sos==4.5.6
speedtest-cli==2.1.3
ssh-import-id==5.11
starlette==0.31.1
systemd-python==235
Twisted==24.3.0
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
uvicorn==0.27.1
uvloop==0.19.0
wadllib==1.3.6
Werkzeug==3.0.1
wheel==0.42.0
wrapt==1.15.0
wsproto==1.2.0
xkit==0.0.0
yarl==1.8.2
zope.interface==6.1
